BUILD: secure product admin views

+++ products/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
+from django.contrib.auth.decorators import login_required
 
 from .models import Product, Category
 from .forms import ProductForm
@@ -66,7 +67,12 @@ def product_detail(request, product_id):
 
     return(render(request, 'products/product_detail.html', context))
 
+@login_required
 def add_product(request):
+    if not request.user.is_superuser: 
+        messages.error(request, 'Sorry, only store admin may perform this action')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -84,7 +90,11 @@ def add_product(request):
             }
     return render(request, template, context)
 
+@login_required
 def edit_product(request, product_id):
+    if not request.user.is_superuser: 
+        messages.error(request, 'Sorry, only store admin may perform this action')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -106,7 +116,11 @@ def edit_product(request, product_id):
 
     return render(request, template,context)
 
+@login_required
 def delete_product(request, product_id):
+    if not request.user.is_superuser: 
+        messages.error(request, 'Sorry, only store admin may perform this action')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted')
